Create Program
This is a program where you can analyze stress of the user by having the data of Max bpm,Avg bpm,Resting bpm,Weight and fat percentage. I have taken a data set from Kaggle for the analysis. I recommend either you collect the data with the following Paraments or Just use this functions if you need to calculate stress.